Synchronized with template_stemlib Stable_2010-11-14 Added glViewport call to Target_resized() in test harness
git-svn-id: http://ludobloom.com/svn/StemLibProjects@720 d291a34c-4c4d-0410-8556-a73de5c6a500

+++ glutshell/trunk/source/glutshell/GLUTShell.c
@@ -88,7 +88,7 @@ double Shell_getCurrentTime() {
 	if (timebaseInfo.denom == 0) {
 		mach_timebase_info(&timebaseInfo);
 	}
-	return mach_absolute_time() * timebaseInfo.numer / timebaseInfo.denom * 0.000000001;
+	return mach_absolute_time() * (double) timebaseInfo.numer / timebaseInfo.denom * 0.000000001;
 	
 #elif defined(WIN32)
 	static LARGE_INTEGER frequency;
